Simplify filtering logic further
Rails should be able to handle merging the two queries fine.

This means that newly, also on range=group (the default),
only_public_data is fetched from the database. Previously, this happened
only on range=layer or range=deep, as noted in #897. But in the commit
message of a95955c it is mentioned that also in the range=group case, we
should at some point use merge to combine the two queries for access
checking and filtering.

Since we ran into issues with range=deep and range=layer, but not with
range=group in #1477 and #1660, this is as good a time as any to unify
the code for all ranges.

This change breaks the "additional columns" / table_displays feature,
but we should fix that feature to add all displayed columns to the
select clause of the query (after doing proper permission checks of
course). This should fix https://help.puzzle.ch/#ticket/zoom/3658